requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.2.12
langchain-community>=0.2.12
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.0.4

# Embeddings + Vector DB
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Google Gemini API
google-generativeai>=0.8.0
